README.MD UPDATED, error logging for empty file path

+++ src/fileManager.ts
@@ -2,10 +2,20 @@
 import { promises as fs } from "fs";
 import * as path from "path";
 
+const NewFileName = new Date().getMilliseconds();
+let fileName = NewFileName + "newFileName";
+console.log(fileName);
+
 const fileManager = {
   createFile: async (filePath: string, content: string): Promise<string> => {
+    if (!filePath) {
+      filePath = fileName;
+    }
+    if (!content) {
+      return "Your can`t create an empty file!!!";
+    }
     try {
-      await fs.writeFile(filePath, content);
+      await fs.writeFile(filePath || fileName, content);
       return "File created successfully";
     } catch (error: any) {
       throw new Error(`Failed to create file: ${error.message}`);
@@ -13,8 +23,11 @@ const fileManager = {
   },
 
   readFile: async (filePath: string): Promise<string> => {
+    if (!filePath) {
+      filePath = fileName;
+    }
     try {
-      const data = await fs.readFile(filePath, "utf8");
+      const data = await fs.readFile(filePath || fileName, "utf8");
       return data;
     } catch (error: any) {
       throw new Error(`Failed to read file: ${error.message}`);
@@ -22,8 +35,14 @@ const fileManager = {
   },
 
   updateFile: async (filePath: string, content: string): Promise<string> => {
+    if (!filePath) {
+      filePath = fileName;
+    }
+    if (!content) {
+      return "Your can`t Update an empty file!!!";
+    }
     try {
-      await fs.writeFile(filePath, content);
+      await fs.writeFile(filePath || fileName, content);
       return "File updated successfully";
     } catch (error: any) {
       throw new Error(`Failed to update file: ${error.message}`);
@@ -31,6 +50,9 @@ const fileManager = {
   },
 
   deleteFile: async (filePath: string): Promise<string> => {
+    if (!filePath) {
+      return "Invalid file name. Try again!";
+    }
     try {
       await fs.unlink(filePath);
       return "File deleted successfully";
